+++ packages/contracts/migrations/index.ts
@@ -1,0 +1,5 @@
+import { deployer } from '../test/utils/deployer';
+
+import { runMigrationsAsync } from './migrate';
+
+runMigrationsAsync(deployer);