+++ packages/subproviders/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.7.0 - _TBD_
 
     * Updated legerco packages. Removed node-hid package as a dependency and make it an optional dependency. It is still used in integration tests but is causing problems for users on Linux distros. (#437)
+    * Add coverage subprovider (#TBD)
 
 ## v0.6.0 - _March 4, 2018_